move NullResultWriter from i_write.h to null_result_writer.h
