+++ src/libs/antares/writer/CMakeLists.txt
@@ -39,8 +39,10 @@ add_library(result_writer
         ensure_queue_started.cpp
         immediate_file_writer.cpp
         in_memory_writer.cpp
+        null_result_writer.cpp
         include/antares/writer/i_table_writer.h
         include/antares/writer/in_memory_writer.h # Public (for tests)
+        include/antares/writer/null_result_writer.h
         include/antares/writer/result_format.h
         include/antares/writer/table_format.h
         include/antares/writer/table_writer_factory.h
