+++ src/libs/antares/writer/include/antares/writer/i_writer.h
@@ -31,14 +31,4 @@ public:
     virtual bool needsTheJobQueue() const = 0;
     virtual void finalize(bool verbose) = 0;
 };
-
-class NullResultWriter: public Solver::IResultWriter
-{
-    void addEntryFromBuffer(const std::filesystem::path&, std::string&) override;
-    void addEntryFromFile(const std::filesystem::path&, const std::filesystem::path&) override;
-    void flush() override;
-    bool needsTheJobQueue() const override;
-    void finalize(bool) override;
-};
-
 } // namespace Antares::Solver
