+++ src/tests/inmemory-study/include/in-memory-study.h
@@ -5,6 +5,7 @@
 #define WIN32_LEAN_AND_MEAN
 #include <limits>
 
+#include <antares/writer/null_result_writer.h>
 #include "antares/io/outputs/SimulationTableCsv.h"
 #include "antares/solver/simulation/ISimulationObserver.h"
 #include "antares/solver/simulation/economy.h"
